refactor scraping code to separate file
